+++ src/Arguments/Values/ValueInterface.php
@@ -30,6 +30,13 @@ interface ValueInterface
     /**
      * @return string
      */
+    public function dumpAsString();
+
+    /**
+     * @deprecated
+     *
+     * @return mixed
+     */
     public function getDump();
 
     public function __toString();