+++ tests/Arguments/Values/ValueTest.php
@@ -28,8 +28,8 @@ class ValueTest extends BaseTestCase
     {
         $this->assertTrue($value->isRealValueReadable());
         $this->assertSame($expectedRealValue, $value->getRealValue());
-        $this->assertSame($value->getDump(), (string)$value);
-        $this->expectOutputString($value->getDump());
+        $this->assertSame($value->dumpAsString(), (string)$value);
+        $this->expectOutputString($value->dumpAsString());
         $value->dump();
     }
 
@@ -63,8 +63,8 @@ class ValueTest extends BaseTestCase
     {
         /** @var Value $restored */
         $restored = \unserialize(\serialize($value));
-        $this->assertSame($value->getDump(), $restored->getDump());
-        $this->assertSame((string)$value, $restored->getDump());
+        $this->assertSame($value->dumpAsString(), $restored->dumpAsString());
+        $this->assertSame((string)$value, $restored->dumpAsString());
         $this->assertSame($value->getRealValue(), $restored->getRealValue());
     }